+++ data/doctors.csv
@@ -1,0 +1,27 @@
+doctorId,firstName,lastName,gender,age,type,mobileNumber,address
+1,John,Doe,M,43,Cardiologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+2,Isabella,Cullen,F,28,Physician,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+3,Rhea,Jain,F,29,Cardiologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+4,Dhananjay,Dey,M,44,Psychatrist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+5,James,Carwin,M,38,Neurologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+6,Konark,Sharma,M,29,Physician,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+7,Mark,Andrew,M,32,Psychatrist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+8,Melinda,Andrew,F,44,Cardiologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+9,Isabella,Carwin,F,29,Psychatrist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+10,Shaun,Murphy,M,45,Psychatrist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+11,Freddie,Johnson,M,29,Neurologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+12,Regina,Phalange,F,29,Physician,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+13,Ken,Adams,M,30,Physician,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+14,Monica,Jain,F,32,ENT,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+15,Ryan,Mallick,M,34,ENT,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+16,Patrick,Stephens,M,39,Neurologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+17,John,Markson,M,30,Dermatologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+18,Vishal,Mallick,M,44,Dentist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+19,Vishal,Mallick,M,49,Dermatologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+20,Kiara,Jain,F,39,Dermatologist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+21,Mark,Johnson,M,32,Dentist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+22,Akshat,Kumar,M,29,Surgeon,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+23,Akshay,Agrawal,M,32,ENT,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+24,William,Chadwick,M,39,Surgeon,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+25,Rishabh,Jain,M,30,Surgeon,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
+26,Navya,Gupta,M,19,Dentist,+1xxxxxxxxxx,ABC-Street` D-Block` E` F
