--- persons/doctors/doctors.cpp
@@ -1,3 +1,0 @@
-#include "doctor.cpp"
-
-vector<doctor> doctors;
