--- persons/doctors/doctors.csv
@@ -1,26 +1,0 @@
-doctorId,firstName,lastName,gender,age,type
-1,John,Doe,M,43,Cardiologist
-2,Isabella,Cullen,F,28,Physician
-3,Rhea,Jain,F,29,Cardiologist
-4,Dhananjay,Dey,M,44,Psychatrist
-5,James,Carwin,M,38,Neurologist
-6,Konark,Sharma,M,29,Physician
-7,Mark,Andrew,M,32,Psychatrist
-8,Melinda,Andrew,F,44,Cardiologist
-9,Isabella,Carwin,F,29,Psychatrist
-10,Shaun,Murphy,M,45,Psychatrist
-11,Freddie,Johnson,M,29,Neurologist
-12,Regina,Phalange,F,29,Physician
-13,Ken,Adams,M,30,Physician
-14,Monica,Jain,F,32,ENT
-15,Ryan,Mallick,M,34,ENT
-16,Patrick,Stephens,M,39,Neurologist
-17,John,Markson,M,30,Dermatologist
-18,Vishal,Mallick,M,44,Dentist
-19,Vishal,Mallick,M,49,Dermatologist
-20,Kiara,Jain,F,39,Dermatologist
-21,Mark,Johnson,M,32,Dentist
-22,Akshat,Kumar,M,29,Surgeon
-23,Akshay,Agrawal,M,32,ENT
-24,William,Chadwick,M,39,Surgeon
-25,Rishabh,Jain,M,30,Surgeon
