--- persons/person.cpp
@@ -1,14 +1,0 @@
-#ifndef PERSON
-#define PERSON
-#include <vector>
-#include <string>
-#include <iostream>
-#include <sstream>
-class person
-{
-protected:
-    string firstName, lastName;
-    char gender;
-    int16_t age = 0;
-};
-#endif // !PERSON
